decomp a few and combine few files
