+++ src/dungeon_ai_attack.c
@@ -13,7 +13,6 @@
 #include "dungeon_action.h"
 #include "dungeon_ai_targeting.h"
 #include "dungeon_ai_targeting_1.h"
-#include "dungeon_ai_targeting_2.h"
 #include "dungeon_capabilities_1.h"
 #include "dungeon_global_data.h"
 #include "dungeon_map_access.h"
