+++ src/dungeon_ai_item_weight.c
@@ -3,7 +3,7 @@
 
 #include "constants/status.h"
 #include "constants/targeting.h"
-#include "dungeon_ai_targeting_2.h"
+#include "dungeon_ai_targeting_1.h"
 #include "dungeon_map_access.h"
 #include "dungeon_pokemon_attributes.h"
 #include "dungeon_util.h"
