+++ src/dungeon_ai_items.c
@@ -9,7 +9,7 @@
 #include "dungeon_ai_attack.h"
 #include "dungeon_ai_item_weight.h"
 #include "dungeon_ai_items.h"
-#include "dungeon_ai_targeting_2.h"
+#include "dungeon_ai_targeting_1.h"
 #include "dungeon_capabilities.h"
 #include "dungeon_capabilities_1.h"
 #include "dungeon_entity.h"
